feat: refactor module-load/run-script into meta.refl
