+++ src/pico/stdlib/extra.c
@@ -8,13 +8,12 @@
 
 #include "pico/stdlib/core.h"
 #include "pico/stdlib/extra.h"
-#include "pico/stdlib/meta.h"
+#include "pico/stdlib/meta/meta.h"
 #include "pico/syntax/concrete.h"
 #include "pico/analysis/abstraction.h"
 #include "pico/codegen/foreign_adapters.h"
 #include "pico/codegen/internal.h"
 
-#include "app/module_load.h"
 
 static jump_buf* m_buf;
 void set_exit_callback(jump_buf* buf) { m_buf = buf; }
@@ -26,7 +25,13 @@ static uint64_t std_perm_allocator;
 static uint64_t std_temp_allocator; 
 
 static Package* current_package;
-void set_current_package(Package* current) { current_package = current; }
+Package* set_current_package(Package* current) {
+    Package* old = current_package;
+    current_package = current;
+    return old;
+}
+
+Package* get_current_package() { return current_package; }
 
 static IStream* current_istream;
 void set_std_istream(IStream* current) { current_istream = current; }
@@ -36,6 +41,10 @@ OStream *set_std_ostream(OStream *current) {
     OStream* old = current_ostream;
     current_ostream = current;
     return old;
+}
+
+OStream *get_std_ostream() {
+    return current_ostream;
 }
 
 Allocator get_std_current_allocator() {
@@ -230,68 +239,6 @@ void build_println_fn(PiType* type, Assembler* ass, Allocator* a, ErrorPoint* po
     CType fn_ctype = mk_fn_ctype(a, 1, "string", mk_string_ctype(a), (CType){.sort = CSVoid});
     convert_c_fn(relic_println_fn, &fn_ctype, type, ass, a, point); 
     delete_c_type(fn_ctype, a);
-}
-
-// C implementation (called from pico!)
-Result load_module_c_fun(String filename) {
-    // (ann load-module String → Unit) : load the module/code located in file! 
-    
-    Allocator* a = get_std_allocator();
-    IStream* sfile = open_file_istream(filename, a);
-    if (!sfile) {
-      return (Result) {
-        .type = Err, .error_message = mv_string("failed to open file!"),
-      };
-    }
-    FormattedOStream* os = mk_formatted_ostream(current_ostream, a);
-
-    Module* parent = get_std_current_module();
-    load_module_from_istream(sfile, os, current_package, parent, a);
-    delete_istream(sfile, a);
-    delete_formatted_ostream(os, a);
-    return (Result) {.type = Ok};
-}
-
-void build_load_module_fun(PiType* type, Assembler* ass, Allocator* a, ErrorPoint* point) {
-    CType result_ctype = mk_struct_ctype(a, 2,
-                                         "type", mk_primint_ctype((CPrimInt){.prim = CLongLong, .is_signed = Unsigned}),
-                                         "error_message", copy_c_type(mk_string_ctype(a), a));
-
-
-    CType fn_ctype = mk_fn_ctype(a, 1, "filename", mk_string_ctype(a), result_ctype);
-
-    convert_c_fn(load_module_c_fun, &fn_ctype, type, ass, a, point); 
-}
-
-Result run_script_c_fun(String filename) {
-    // (ann load-module String → Unit) : load the module/code located in file! 
-    
-    Allocator* a = get_std_allocator();
-    IStream* sfile = open_file_istream(filename, a);
-    if (!sfile) {
-      return (Result) {
-        .type = Err, .error_message = mv_string("failed to open file!"),
-      };
-    }
-    Module* current_module = get_std_current_module();
-    FormattedOStream* os = mk_formatted_ostream(current_ostream, a);
-    run_script_from_istream(sfile, os, current_module, a);
-    delete_istream(sfile, a);
-    delete_formatted_ostream(os, a);
-    return (Result) {.type = Ok};
-}
-
-void build_run_script_fun(PiType* type, Assembler* ass, Allocator* a, ErrorPoint* point) {
-    // Proc type
-    // TODO: this does not accurately represent the result type!
-    CType result_ctype = mk_struct_ctype(a, 2,
-                                         "type", mk_primint_ctype((CPrimInt){.prim = CLongLong, .is_signed = Unsigned}),
-                                         "error_message", copy_c_type(mk_string_ctype(a), a));
-
-
-    CType fn_ctype = mk_fn_ctype(a, 1, "filename", mk_string_ctype(a), result_ctype);
-
-    convert_c_fn(run_script_c_fun, &fn_ctype, type, ass, a, point); 
 }
 
 RawTree atom_symbol(const char *str) {
@@ -908,24 +855,6 @@ void add_extra_module(Assembler* ass, Package* base, Allocator* default_allocato
 
     build_println_fn(typep, ass, a, &point);
     sym = string_to_symbol(mv_string("print-ln"));
-    fn_segments.code = get_instructions(ass);
-    prepped = prep_target(module, fn_segments, ass, NULL);
-    add_def(module, sym, *typep, &prepped.code.data, prepped, NULL);
-    clear_assembler(ass);
-
-    // load-module : Proc [String] Unit
-    typep = mk_proc_type(a, 1, mk_string_type(a), mk_enum_type(a, 2, "Ok", 0, "Err", 1, mk_string_type(a)));
-    build_load_module_fun(typep, ass, a, &point);
-    sym = string_to_symbol(mv_string("load-module"));
-    fn_segments.code = get_instructions(ass);
-    prepped = prep_target(module, fn_segments, ass, NULL);
-    add_def(module, sym, *typep, &prepped.code.data, prepped, NULL);
-    clear_assembler(ass);
-
-    // run-script : Proc [String] Result
-    typep = mk_proc_type(a, 1, mk_string_type(a), mk_enum_type(a, 2, "Ok", 0, "Err", 1, mk_string_type(a)));
-    build_run_script_fun(typep, ass, a, &point);
-    sym = string_to_symbol(mv_string("run-script"));
     fn_segments.code = get_instructions(ass);
     prepped = prep_target(module, fn_segments, ass, NULL);
     add_def(module, sym, *typep, &prepped.code.data, prepped, NULL);
