+++ src/pico/stdlib/stdlib.c
@@ -7,7 +7,7 @@
 #include "pico/stdlib/platform/platform.h"
 #include "pico/stdlib/num.h"
 #include "pico/stdlib/extra.h"
-#include "pico/stdlib/meta.h"
+#include "pico/stdlib/meta/meta.h"
 #include "pico/stdlib/debug.h"
 #include "pico/stdlib/foreign.h"
 #include "pico/stdlib/user.h"
