+++ test/src/test_pico/eval/eval.c
@@ -1,3 +1,4 @@
+#include "platform/signals.h"
 #include "platform/memory/executable.h"
 #include "platform/memory/arena.h"
 
@@ -32,8 +33,13 @@ void run_pico_eval_tests(TestLog* log, Allocator* a) {
         .imports = imports,
         .exports = exports,
     };
+
+    ErrorPoint point;
+    if (catch_error(point)) {
+        panic(mv_string("Error in tests: test_pico/eval/eval.c"));
+    }
     Module* module = mk_module(header, base, NULL, a);
-    Environment* env = env_from_module(module, a);
+    Environment* env = env_from_module(module, &point, a);
     delete_module_header(header);
 
     if (suite_start(log, mv_string("literals"))) {
